chore(transforms): Refactor output types into sub-module
